Added Menu Driven Functionality
The Program now has fair share of Validation and Menu Driven Functionality. Modifications has been done to the Code to achieve that.

Also the login_logoutMod.py has been edited to return a messege "User already logges out", if the user Specified exists and is already logged out of the System.

And, the file menu.py has been edited with an extra Option, for quitting the Main Menu.

Commiter ~ @iamapurba2003
Modifier ~ @iamapurba2003